% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontgis-package.R
\docType{package}
\name{fontgis-package}
\alias{fontgis}
\alias{fontgis-package}
\title{fontgis: A simple way to use GIS icons in R for spatial visualisation and analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Insert 'Font-GIS' icon sets into 'R Markdown' documents, 'Quarto' files, and 'Shiny' applications in a simple and flexible way. The icons, designed for GIS and spatial analysis themes, can be integrated as 'SVG' elements or HTML '<i>' tags. They can also be exported as 'PNG' images for use in contexts requiring raster graphics.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://geografo.pe/fontgis/}
}

}
\author{
\strong{Maintainer}: antony barja \email{geografope@gmail.com}

}
\keyword{internal}
