% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_i.R
\name{gis_i}
\alias{gis_i}
\title{Renderiza i tag GIS como SVG en el Viewer (usando alias)}
\usage{
gis_i(name, width = "1em", height = "1em", fill = "currentColor")
}
\arguments{
\item{name}{Alias amigable del icono (por ejemplo: "north-arrow")}

\item{width}{Ancho (por ejemplo "100px", "1em")}

\item{height}{Alto}

\item{fill}{Color de relleno del icono}
}
\value{
HTML visual en Viewer.
}
\description{
Renderiza i tag GIS como SVG en el Viewer (usando alias)
}
