% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_icon.R
\name{gis_icon}
\alias{gis_icon}
\title{Renderiza un icono GIS como SVG en el Viewer (usando alias)}
\usage{
gis_icon(name, width = "100px", height = "100px", fill = "black")
}
\arguments{
\item{name}{Alias amigable del icono (por ejemplo: "north-arrow")}

\item{width}{Ancho (por ejemplo "100px", "1em")}

\item{height}{Alto}

\item{fill}{Color de relleno del icono}
}
\value{
HTML visual en Viewer.
}
\description{
Renderiza un icono GIS como SVG en el Viewer (usando alias)
}
