% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_icon.R
\name{gis_icons}
\alias{gis_icons}
\title{Listar todos los iconos disponibles con alias}
\usage{
gis_icons()
}
\value{
Un data frame con columnas: alias, name, category
}
\description{
Listar todos los iconos disponibles con alias
}
