% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_png.R
\name{gis_png}
\alias{gis_png}
\title{Exporta un ícono GIS a PNG}
\usage{
gis_png(
  name,
  file = tempfile(fileext = ".png"),
  width = 100,
  height = 100,
  fill = "black"
)
}
\arguments{
\item{name}{Alias del ícono}

\item{file}{Ruta de salida (.png)}

\item{width}{Tamaño en px}

\item{height}{Tamaño en px}

\item{fill}{Color del ícono}
}
\value{
Ruta del archivo PNG generado
}
\description{
Exporta un ícono GIS a PNG
}
